requirements.txt
asgiref                       3.6.0
Django                        4.1.5
django-cors-headers           3.13.0
django-dotenv                 1.4.2
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
pip                           21.3.1
PyJWT                         2.6.0
pytz                          2022.7.1
setuptools                    60.2.0
sqlparse                      0.4.3
wheel                         0.37.1
